chore: reorganizes files based on client/server structure
